+++ tests/helpers/lineplot-loader.js
@@ -95,4 +95,5 @@ module.exports = {
   formatX: lineplotHelpers.formatX,
   computeSeries: lineplotHelpers.computeSeries,
   computePerXStats: lineplotHelpers.computePerXStats,
+  lineDashArray: lineplotHelpers.lineDashArray,
 };